\name{BIOGRID.PSIMI25}
\alias{BIOGRID.PSIMI25}
\alias{DIP.PSIMI25}
\alias{HPRD.PSIMI25}
\alias{INTACT.PSIMI25}
\alias{MINT.PSIMI25}
\alias{MIPS.PSIMI25}
\alias{MATRIXDB.PSIMI25}
\alias{CORUM.PSIMI25}
\alias{MPACT.PSIMI25}
\docType{data}
\title{ Supported PSI-MI 2.5 XML data sources}
\description{
  Objects representing supported PSI-MI 2.5 XML data sources. They are
  implicitly used in interaction and/or complex parsers. The supporting
  list will grow as soon as there are new PSI-MI 2.5 compatible data
  sources available.

  For usage see examples of \code{\link{parsePsimi25Interaction}},
  \code{\link{parsePsimi25Complex}} and \code{\link{psimi25XML2Graph}}.
}
\details{
  If a new repository is added manually by the user, please read the
  notes in \code{\link{psimi25Source-class}} before coding
}
\references{
 \describe{
  \item{PSI-MI XML v2.5 data exchange
    format}{\url{http://www.psidev.info/index.php?q=node/60}}
  \item{BioGRID}{\url{http://www.thebiogrid.com}}
  \item{DIP}{\url{http://dip.doe-mbi.ucla.edu/}}
  \item{HPRD}{\url{http://www.hprd.org/}}
  \item{IntAct}{\url{http://www.ebi.ac.uk/intact}}
  \item{MINT}{\url{http://mint.bio.uniroma2.it/mint}}
  \item{MIPS/CORUM}{\url{http://mips.gsf.de/}}
  \item{MPact}{\url{http://mips.gsf.de/genre/proj/mpact}}
 }
}
\keyword{models}
