\name{interactorInfo-methods}
\docType{methods}
\alias{interactorInfo-methods}
\alias{interactorInfo,list-method}
\alias{interactorInfo,psimi25ComplexEntry-method}
\alias{interactorInfo,psimi25Graph-method}
\alias{interactorInfo,psimi25InteractionEntry-method}
\title{ Methods for Function interactorInfo in Package `RpsiXML'}
\description{
  See interactorInfo 
}
\section{Methods}{
\describe{

\item{x = "list"}{ A list of \code{\link{psimi25Interactor-class}} objects}

\item{x = "psimi25ComplexEntry"}{ Object of
  \code{\link{psimi25ComplexEntry-class}}, parsed from PSI-MI 2.5 complex files}

\item{x = "psimi25Graph"}{Object of \code{\link{psimi25Graph-class}}, parsed
  by \code{\link{psimi25XML2Graph}}}

  by \code{\link{psimi25XML2Graph}}

\item{x = "psimi25InteractionEntry"}{ Object of
  \code{\link{psimi25InteractionEntry-class}}, parsed from PSI-MI 2.5
  interaction files }
}
}
\keyword{methods}

