\name{interactors-methods}
\docType{methods}
\alias{interactors-methods}
\alias{interactors,psimi25Graph-method}
\alias{interactors,psimi25Hypergraph-method}
\title{Methods for Function interactors}
\description{
  Methods to extract interactors from various data structures
}
\section{Methods}{
\describe{
\item{x = "psimi25ComplexEntry"}{ Object of
  \code{\link{psimi25ComplexEntry-class}}, parsed from PSI-MI 2.5 complex files}

\item{x = "psimi25Graph"}{Object of \code{\link{psimi25Graph-class}}, parsed
  by \code{\link{psimi25XML2Graph}}}

\item{x = "psimi25Hypergraph"}{ Object of \code{\link{psimi25Hypergraph-class}}, parsed
  by \code{\link{psimi25XML2Graph}}}

\item{x = "psimi25InteractionEntry"}{ Object of
  \code{\link{psimi25InteractionEntry-class}}, parsed from PSI-MI 2.5
  interaction files }

}}
\keyword{methods}
