\name{numInteractions-methods}
\docType{methods}
\alias{numInteractions}
\alias{numInteractions-methods}
\alias{numInteractions,psimi25InteractionEntry-method}
\title{ Get interaction number of the given object}
\description{
  Get the interaction number of the given object
}
\section{Methods}{
\describe{

\item{x = "psimi25InteractionEntry"}{ Returns the interactio number of
  the interactionEntry}
}}
\keyword{methods}

