\name{numInteractors-methods}
\docType{methods}
\alias{numInteractors}
\alias{numInteractors-methods}
\alias{numInteractors,psimi25Hypergraph-method}
\alias{numInteractors,psimi25InteractionEntry-method}
\title{ Get interactor number of the given item }
\description{
  Get the number of interactors in the given item
}
\section{Methods}{
\describe{

\item{x = "psimi25Hypergraph"}{ Get the protein number in the hypergraph
object}

\item{x = "psimi25InteractionEntry"}{ Get the interactor number in the entry }
}}
\keyword{methods}
