\name{sourceDb-methods}
\docType{methods}
\alias{sourceDb-methods}
\alias{sourceDb,psimi25Complex-method}
\alias{sourceDb,psimi25Interaction-method}
\alias{sourceDb,psimi25Interactor-method}
\alias{sourceDb,psimi25Source-method}

\title{ Get source database name}
\description{
  The method returns source database name of the given parameter
}
\section{Methods}{
\describe{

\item{x = "psimi25Complex"}{ An object of \code{\link{psimi25Complex-class}} }

\item{x = "psimi25Interaction"}{ An object of \code{\link{psimi25Interaction-class}} }

\item{x = "psimi25Interactor"}{ An object of \code{\link{psimi25Interactor-class}} }

\item{x = "psimi25Source"}{ An object of \code{\link{psimi25Source-class}} }
}}
\keyword{methods}
\keyword{models}
