\name{sourceDb}
\alias{sourceDb}

\title{ Extracts the source database of the given parameter }
\description{
  Returns the name of the source database of the given object
}
\usage{
sourceDb(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ see \code{\link{sourceDb-methods}} }
}
\value{
  A character string presenting the source database
  }
\author{ Jitao David Zhang <j.zhang@dkfz.de>, Tony Chiang <tchiang@ebi.ac.uk> }
\seealso{see \code{\link{sourceDb-methods}} }
\keyword{ models}
