\name{sourceId-methods}
\docType{methods}
\alias{sourceId-methods}
\alias{sourceId,psimi25Complex-method}
\alias{sourceId,psimi25Interaction-method}
\alias{sourceId,psimi25Interactor-method}
\title{ Extracting ID in the source database}
\description{
  The method extracts source database IDs of the given parameter
}
\section{Methods}{
\describe{

\item{x = "psimi25Complex"}{ An object of \code{\link{psimi25Complex-class}} }

\item{x = "psimi25Interaction"}{ An object of \code{\link{psimi25Interaction-class}} }

\item{x = "psimi25Interactor"}{ An object of \code{\link{psimi25Interactor-class}} }
}}
\keyword{methods}
\keyword{models}
