\name{sourceId}
\alias{sourceId}
\title{ Extracts the source database ID of the given paramter}
\description{
  see \code{\link{sourceDb-methods}}
}
\usage{
sourceId(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ \code{\link{sourceDb-methods}} }
}
\value{
  A character string presenting ID in the source database
}
\author{ Jitao David Zhang <j.zhang@dkfz.de>, Tony Chiang <tchiang@ebi.ac.uk> }
\seealso{see \code{\link{sourceId-methods}} }
\keyword{ models}