\name{taxId-methods}
\docType{methods}
\alias{taxId-methods}
\alias{taxId,psimi25Complex-method}
\alias{taxId,psimi25Interactor-method}
\title{ Methods for function taxId}
\description{
  Returns the NCBI Taxonomy ID of the given object
}
\section{Methods}{
\describe{

  \item{x = "psimi25Complex"}{ An \code{\link{psimi25Complex-class}}
    object }
  \item{x = "psimi25Interactor"}{ An \code{\link{psimi25Interactor-class}}
    object }
  
\item{x = "psimi25InteractionEntry"}{ An object of
  \code{\link{psimi25InteractionEntry-class}} }
}}
\keyword{methods}

