\name{uniprot-methods}
\docType{methods}
\alias{uniprot-methods}
\alias{uniprot,psimi25Interactor-method}
\title{ Methods for Function uniprot in Package `RpsiXML'}
\description{
  if the given parameter is an object of
  \code{\link{psimi25Interactor-class}}, then the UniProt ID of this
  protein is returned. In case it is an object of
  \code{\link{psimi25Source-class}}, the character string representing
  the path to UniProt is returned.
}
\section{Methods}{
\describe{

\item{x = "psimi25Interactor"}{ An object of
  \code{\link{psimi25Interactor-class}}  }

\item{x = "psimi25Source"}{ An object of
  \code{\link{psimi25Source-class}}
}
}}
\keyword{methods}

