\name{psimi25Hypergraph2GraphNEL}
\Rdversion{1.1}
\alias{psimi25Hypergraph2GraphNEL}
\title{Convert psimi25Hypergraph to graphNEL}
\description{
  Convert a \code{psimi25Hypergraph} object to \code{graphNEL} object,
  for the purpose of modelling, visualization, etc.
    
  NA nodes will be first removed from the hypergraph, and then the
  hypergraph is converted to the graphNEL object
}
\usage{
psimi25Hypergraph2GraphNEL(x)
}
\arguments{
  \item{x}{An object of the class \code{psimi25Hypergraph}}
}
\value{
  An object of the class \code{graphNEL}
}
\author{
  Jitao David Zhang <jitao_david.zhang@roche.com>
}
\examples{
## to be implemented
}
