\name{sourceId-methods}
\docType{methods}
\alias{sourceDb}
\alias{sourceDb<-}
\alias{sourceId}
\alias{sourceId<-}
\alias{sourceId-methods}
\alias{sourceDb-methods}
\alias{sourceId,sourceDbAndId-method}
\alias{sourceDb,sourceDbAndId-method}
\alias{sourceDb,psimi25Source-method}
\alias{"sourceId<-"-methods}
\alias{"sourceDb<-"-methods}
\alias{sourceId<-,sourceDbAndId,character-method}
\alias{sourceDb<-,sourceDbAndId,character-method}
\title{ Extract or set source database name or ID in the source database}
\description{
  The methods extracts or sets the source database where the object comes from,
  or its identifier there.
}
\section{Methods}{
\describe{
\item{x = "sourceDbAndId"}{ An object of \code{sourceDbAndId} (internal), or
one of its subclasses}
}}
\keyword{methods}
\keyword{models}
