\name{taxId-methods}
\docType{methods}
\alias{taxId}
\alias{taxId-methods}
\alias{taxId,organismTaxIdAndName-method}
\alias{taxId<-}
\alias{taxId<-,organismTaxIdAndName,character-method}
\alias{organismName}
\alias{organismName-methods}
\alias{organismName,organismTaxIdAndName-method}
\alias{organismName<-}
\alias{organismName<-,organismTaxIdAndName,character-method}
\title{Get or Set the NCBI Taxonomy ID or Organism Name}
\description{
  Get or set the NCBI taxonomy ID or organism name
}
\section{Methods}{
\describe{
  \item{x = "organismTaxIdAndName"}{ An \code{organismTaxIdAndName}
    object}
}}
\keyword{methods}

